Move readSms from onResume to onCreate
After saving a file, the loading indicator would spin because readSms
has been called again. This can mislead the user into believing that
the loading is for the file saving the file. Moving readSms to onCreate
prevents this behaviour. SwipeRefreshLayout is now used to allow the
user to refresh the data. The original loading indicator has also been
removed, in favour of the refreshLayout's.